TG-23: (WIP) fixed routes

+++ src/controllers/projects.js
@@ -2,7 +2,7 @@ import Project from '../models/Projects';
 
 const getProject = async (req, res) => {
   try {
-    const allProjects = await Project.find(req.body).populate('id', 'name');
+    const allProjects = await Project.find(req.body);
     if (allProjects.length === 0) {
       return res.status(404).json({
         message: 'Not found',
@@ -52,20 +52,11 @@ const addProject = async (req, res) => {
   try {
     const project = new Project({
       name: req.body.name,
+      isActive: req.body.isActive,
       description: req.body.description,
+      client: req.body.client,
       startDate: req.body.startDate,
       endDate: req.body.endDate,
-      clientName: req.body.clientName,
-      active: req.body.active,
-      devRate: req.body.devRate,
-      qaRate: req.body.qaRate,
-      pmRate: req.body.pmRate,
-      tlRate: req.body.tlRate,
-      devs: req.body.devs,
-      qas: req.body.qas,
-      projectManager: req.body.projectManager,
-      techLeader: req.body.techLeader,
-      admin: req.body.admin,
     });
     const result = await project.save();
     return res.status(201).json({
@@ -99,7 +90,7 @@ const updateProject = async (req, res) => {
     if (!result) {
       return res.status(404).json({
         message: 'Project not found',
-        data: undefined,
+        data: null,
         error: true,
       });
     }
@@ -119,26 +110,29 @@ const updateProject = async (req, res) => {
 
 const deleteProject = async (req, res) => {
   try {
+    if (!req.params) {
+      return res.status(400).json({
+        msg: 'Missing id parameter',
+      });
+    }
     const result = await Project.findByIdAndDelete(req.params.id);
     if (!result) {
       return res.status(404).json({
         message: 'Project not found',
-        data: undefined,
+        data: null,
         error: true,
       });
     }
-    return res.status(200).json({
+    return res.status(204).json({
       message: 'Project has been deleted',
       data: result,
       error: false,
-
     });
   } catch (error) {
     return res.json({
       message: error.message,
       data: undefined,
       error: true,
-
     });
   }
 };
